requirements.txt
# Core (required)
click>=8.0
rich>=13.0
requests>=2.28

# AWS
boto3>=1.28
botocore>=1.31

# GCP
google-auth>=2.0
google-api-python-client>=2.0
cryptography>=41.0

# Azure
azure-identity>=1.14
azure-storage-blob>=12.0
azure-mgmt-authorization>=3.0

# Development
pytest>=7.0
pytest-mock>=3.0
